gromacs gpu test runs
