+++ configuration/cirrus_settings.py
@@ -26,7 +26,8 @@ site_configuration = {
                     "max_jobs": 16,
                     "environs": ["gnu", "intel"],
                     "resources": [
-                        {"name": "qos", "options": ["--qos={qos}"]},
+                        {"name": "qos", "options": ["--qos={qos}"],
+                        },
                     ],
                 },
                 {
@@ -42,7 +43,8 @@ site_configuration = {
                     "max_jobs": 4,
                     "environs": ["nvidia-mpi"],
                     "resources": [
-                        {"name": "qos", "options": ["--qos={gpu}"]},
+                        {"name": "qos", "options": ["--qos={qos_id}"]},
+                        {"name": "gpu", "options": ["--gres=gpu:{num_gpus_per_node}"]},
                     ],
                 },
             ],
